Only Double needs to be supported
Double is an `IntegerLiteralConvertible`. Using an Integer literal will
convert it to `Double` so we don't need to have different methods on
every different number type. This also fixes 32 bits issues.

Note that the time methods are still only `Int`. This is because these
methods are essentially wrappers around `NSDateComponents` which only
accepts `Int`s

Closes #2